rendered list of doctors
