+++ src/components/doctor/ListDoctor.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import ShowDoctor from './ShowDoctor';
+
+function ListDoctor(value) {}
+
+export default ListDoctor;
