+++ src/components/sidemenu/SideMenu.jsx
@@ -1,22 +1,1 @@
 import React from 'react';
-
-class SideMenu extends React.Component {
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return (
-      <div>
-        <div>
-          <img src="" alt="logo" />
-        </div>
-        <ul>
-          <NavLink to=""></NavLink>
-        </ul>
-      </div>
-    );
-  }
-}
-
-export default SideMenu;
